Reorder files in directories
