+++ util/webxr/MediaLayerManager/index.js
@@ -1,0 +1,6 @@
+import MediaLayerManager from "./MediaLayerManager";
+import GlassLayer from "./GlassLayer";
+import MediaLayer from "./MediaLayer";
+
+export { GlassLayer, MediaLayer };
+export default MediaLayerManager;
